refactor(backend): decompose courses router and course_service
Both courses.py files topped 500 lines and mixed several unrelated
concerns. Turn each into a package of focused modules.

app/api/v1/courses/
  __init__.py    re-export `router` + side-effect imports
  _router.py     shared APIRouter(prefix="/courses")
  catalog.py     GET /, /my, /my/trash, /{id}, /{id}/modules/{id}
  crud.py        POST /, PUT /{id}, DELETE /{id}, clone/restore/permanent
  modules.py     POST /{id}/modules + PUT/DELETE /{id}/modules/{id}
  chapters.py    POST / PUT / DELETE /{id}/modules/{id}/chapters[/{id}]
  enrollment.py  GET /{id}/enrollment-status + POST /{id}/enroll
                 (plus _enforce_cohort_gates helper extracted from
                  the old inline enroll() body)

app/services/course_service/
  __init__.py    re-exports the public API — callers keep using
                 ``from app.services.course_service import X`` unchanged
  _queries.py    _COURSE_TREE, get_courses / get_course /
                 get_teacher_courses / get_module / get_chapter
  _courses.py    create / update / delete / restore / permanently_delete
  _modules.py    create / update / delete module (+ _next_module_order)
  _chapters.py   create / update / delete chapter (+ _next_chapter_order)
  _enrollment.py enroll_user_in_course, get_user_courses,
                 sync_enrollment_progress
  _clone.py      clone_course (deep copy)

Public surface is unchanged:
  - Same 19 endpoints, same paths, same response models.
  - `from app.api.v1 import courses; courses.router` still works.
  - `from app.services.course_service import <name>` still works for
    every previously-exported symbol (verified via an import smoke
    test covering 20 symbols).

All 396 tests pass; ruff/mypy clean on app/.

Made-with: Cursor
